requirements.txt
fastapi
pydantic
sqlalchemy
load_dotenv
alembic
psycopg2-binary
uvicorn
httpx
pydantic[email]

pyjwt
pwdlib[argon2]